requirements.txt
numpy
pandas
matplotlib
seaborn
shap
joblib
scikit-learn
xgboost
pyyaml
